Add files via upload
I replicated the frontend of an existing website using HTML and Tailwind CSS, utilizing Tailwind CSS's utility-first approach to efficiently style elements and closely mimic the design and layout of the original site. By dissecting the structure and styling nuances of the target website, I honed my skills in HTML for markup and Tailwind CSS for rapid and responsive styling. This project not only reinforced my understanding of frontend development principles but also showcased my ability to recreate complex designs with precision and efficiency.